browser/components/feeds/test/test_bug368464.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=368464
-->
<head>
  <title>Test that RSS 0.90 isn't sniffed</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=368464">Mozilla Bug 368464</a>
<p id="display"><iframe id="testFrame" src="bug368464-data.xml"></iframe></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 368464 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalBrowserRead");
  ok($("testFrame").contentDocument.documentElement.id != "feedHandler",
     "RSS 0.90 shouldn't be sniffed as a feed");
});
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
